# server/src/docs/swagger.yaml
definitions:
...
  models.File:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      id:
        type: integer
      object_key:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Persona:
    properties:
      created_at:
        type: string
      id:
        type: integer
      persona_description:
        type: string
      persona_title:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        type: string
    required:
    - persona_description
    - persona_title
    type: object
  models.Profile:
    properties:
      completed_onboarding_response:
        type: boolean
      created_at:
        type: string
      date_of_birth:
        type: string
      id:
        type: integer
      name:
        type: string
      onboarding_response:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - date_of_birth
    - name
    - phone_number
    - user_id
    type: object
...
  models.Tag:
    properties:
      created_at:
        type: string
      files:
        items:
          $ref: '#/definitions/models.File'
        type: array
      id:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      id:
        type: integer
      personas:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      task_description:
        type: string
      task_name:
        type: string
      updated_at:
        type: string
    required:
    - task_description
    - task_name
    type: object
...
info:
  contact:
    email: oduneye.d@northeastern.edu and dupuguntla@northeastern.edu
    name: David Oduneye and Akshay Dupuguntla
  description: Backend Server for Legacy App
  title: Legacy API
  version: "1.0"
paths:
...
  /api/personas/:
    get:
      description: Returns all personas
      operationId: get-all-personas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Persona'
            type: array
        "404":
          description: Failed to fetch personas
          schema:
            type: string
      summary: Gets all personas
      tags:
      - persona
    post:
      consumes:
      - application/json
      description: Creates a persona
      operationId: create-persona
      parameters:
      - description: Persona
        in: body
        name: persona
        required: true
        schema:
          $ref: '#/definitions/models.Persona'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Persona'
        "400":
          description: Failed to create persona
          schema:
            type: string
      summary: Creates a persona
      tags:
      - persona
  /api/personas/{pid}:
    delete:
      description: Deletes a persona
      operationId: delete-persona
      parameters:
      - description: PersonaID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Persona successfully deleted
          schema:
            type: string
        "404":
          description: Failed to delete persona
          schema:
            type: string
      summary: Deletes a persona
      tags:
      - persona
    get:
      description: Returns a persona from persona id
      operationId: get-persona
      parameters:
      - description: PersonaID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Persona'
        "404":
          description: Failed to fetch persona
          schema:
            type: string
      summary: Gets a persona from persona id
      tags:
      - persona
    patch:
      consumes:
      - application/json
      description: Updates a persona
      operationId: update-persona
      parameters:
      - description: PersonaID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Persona'
        "400":
          description: Failed to update persona
          schema:
            type: string
      summary: Updates a persona
      tags:
      - persona
  /api/personas/{pid}/tasks:
    get:
      description: Returns all tasks for a persona
      operationId: get-persona-tasks
      parameters:
      - description: PersonaID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: Failed to fetch persona tasks
          schema:
            type: string
      summary: Gets all tasks for a persona
      tags:
      - persona
  /api/profiles/:
    get:
      description: Returns all profiles
      operationId: get-all-profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Profile'
            type: array
        "404":
          description: Failed to fetch profiles
          schema:
            type: string
      summary: Gets all profiles
      tags:
      - profile
    post:
      consumes:
      - application/json
      description: Creates a profile
      operationId: create-profile
      parameters:
      - description: Profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: Failed to create profile
          schema:
            type: string
      summary: Creates a profile
      tags:
      - profile
  /api/profiles/{pid}:
    delete:
      description: Deletes a profile
      operationId: delete-profile
      parameters:
      - description: ProfileID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted profile
          schema:
            type: string
        "404":
          description: Failed to delete profile
          schema:
            type: string
      summary: Deletes a profile
      tags:
      - profile
    get:
      description: Returns a profile from profile id
      operationId: get-profile
      parameters:
      - description: ProfileID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
        "404":
          description: Failed to fetch profile
          schema:
            type: string
      summary: Gets a profile from profile id
      tags:
      - profile
    patch:
      consumes:
      - application/json
      description: Updates a profile
      operationId: update-profile
      parameters:
      - description: ProfileID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: Failed to update profile
          schema:
            type: string
      summary: Updates a profile
      tags:
      - profile
  /api/profiles/response/{pid}:
    patch:
      consumes:
      - application/json
      description: Inserts an onboarding response
      operationId: insert-onboarding-response
      parameters:
      - description: ProfileID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: Failed to insert onboarding response
          schema:
            type: string
      summary: Inserts an onboarding response
      tags:
      - profile
...
swagger: "2.0"